COMP: fix all warnings on Mac 10.5

+++ Code/IO/GoImage/GoImageProcessor.cxx
@@ -404,7 +404,7 @@ getVisibleImages()
   boost::tuples::tie( it0, it1 ) =
     m_MegaImageContainer.get< Visibility >().equal_range(true);
 
-  size_t N(0);
+  int N(0);
 
   std::vector< IteratorType > tempVector;
